methods/requirements.txt
rdflib
python-Levenshtein
